fix(sdk): compaction stops deleting the turn it is compacting
The recent-window boundary was snapped FORWARD to keep a tool pair whole, but
that walk skips leading `tool` messages with no stop short of the end of the
transcript. When the whole suffix from the naive boundary is `tool` messages —
one assistant turn fanning out at least `keepRecentMessages` parallel calls,
measured at the start of the next iteration, which is exactly where the check
runs — the boundary landed on `messages.length`, the recent window came back
empty, and the rebuilt transcript held no non-system message at all. The model
was then asked to answer a conversation whose last turn, including the user's
own message, had been removed. The older-message floor guard cannot catch it: in
that shape the older half is the whole transcript.

The boundary is now the largest safe one AT OR BELOW naive, so a pass never
removes more than the naive cut would and at least `keepRecentMessages` original
messages survive verbatim. With no safe boundary the pass is skipped and the
transcript is untouched — one iteration of context headroom is cheaper than the
live turn, and the condition clears itself as soon as the next assistant message
moves the boundary past the tool block. It also stops the leading system prompts
being duplicated into the recent window when naive lands inside the system
prefix. `findSafeTrimIndex` is unchanged and reused as the safety predicate, so
its other (public) callers keep their behaviour.

An empty verification reply is now treated like `COMPLETE`. A truncated turn, a
refusal or an exhausted `llmVerificationMaxTokens` yields an empty string, which
fell through to the append path and stamped a bare `## LLM Verification
Additions` heading with nothing under it — an empty promise that then rode in
every subsequent system prompt for the rest of the run.

Every compaction count is `.int().positive()` instead of `.positive()`. Zod's
base number check rejects only non-numbers and NaN, so `Infinity` and fractions
both parsed. `Infinity` was the dangerous one: it disarms the budget it guards
rather than failing, so `convoTextBudget: Infinity` made `truncateMessages` a
no-op and the whole older history was pasted into the verification prompt. No
clamp was added inside `truncateMessages` — a budget that quietly repairs an
illegal value hides the caller's bug, and the schema is where the boundary
belongs. A host that hand-builds a `CompactionConfig` still bypasses it; that is
stated in the test rather than papered over.

Proven red first: 19 failing assertions across the three new suites (4 on the
cut, 3 on the verifier, 12 on the schema), including "never leaves a system-only
transcript" failing with zero non-system messages. 169/169 green after, across
the compaction, config and runtime/query suites.

+++ packages/sdk/src/config/runtime.ts
@@ -26,20 +26,29 @@ export const CompactionConfigSchema = z.object({
 	 * `tokenBudget` unlimited (0) set this so compaction fires on
 	 * window-pressure instead of being a silent no-op.
 	 */
-	contextWindowTokens: z.number().positive().optional(),
+	// Every field below counts something — tokens, messages, characters,
+	// sentences. `z.number().positive()` was the wrong validator for all of them:
+	// Zod's base number check rejects only non-numbers and NaN, so `Infinity` and
+	// `0.5` both passed. `Infinity` is the dangerous one, because it turns a bound
+	// into a no-op rather than an error — `convoTextBudget: Infinity` makes the
+	// verifier's `charCount > budget` test never true, so `truncateMessages` stops
+	// truncating and the WHOLE history is pasted into the verification prompt.
+	// A fraction is merely nonsense (`keepRecentMessages: 2.5`). `.int()` rejects
+	// both: `Number.isInteger` is false for Infinity and for any fraction.
+	contextWindowTokens: z.number().int().positive().optional(),
 	triggerThreshold: z.number().min(0).max(1).default(0.7),
 	resetThreshold: z.number().min(0).max(1).default(0.4),
-	keepRecentMessages: z.number().positive().default(4),
-	maxToolResults: z.number().positive().default(30),
-	maxListSize: z.number().positive().default(25),
+	keepRecentMessages: z.number().int().positive().default(4),
+	maxToolResults: z.number().int().positive().default(30),
+	maxListSize: z.number().int().positive().default(25),
 	llmVerification: z.boolean().default(true),
-	llmVerificationMaxTokens: z.number().positive().default(2048),
-	richStateThreshold: z.number().positive().default(15),
-	convoTextBudget: z.number().positive().default(12_000),
-	maxSentencesPerTurn: z.number().positive().default(5),
-	maxCharsPerNote: z.number().positive().default(500),
-	maxCharsPerRequirement: z.number().positive().default(300),
-	maxCharsPerTask: z.number().positive().default(400),
+	llmVerificationMaxTokens: z.number().int().positive().default(2048),
+	richStateThreshold: z.number().int().positive().default(15),
+	convoTextBudget: z.number().int().positive().default(12_000),
+	maxSentencesPerTurn: z.number().int().positive().default(5),
+	maxCharsPerNote: z.number().int().positive().default(500),
+	maxCharsPerRequirement: z.number().int().positive().default(300),
+	maxCharsPerTask: z.number().int().positive().default(400),
 })
 
 export type CompactionConfig = z.infer<typeof CompactionConfigSchema>
